% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blue_cliff_record}
\alias{blue_cliff_record}
\title{The Blue Cliff Record}
\format{A \code{tbl} with four variables: \code{collection}, \code{case},
  \code{type}, and \code{text}. \code{collection} is the title of the text,
  "Blue Cliff Record". \code{case} indicates the koan case number.
  \code{type} indicates the text type, one of: "title", the title of the
  case, "main_case", the main case of the koan, or "commentary", commentary
  on the koan by Setcho.}
\source{
http://4x13.net/buddhism/koan_bluecliff.pdf
}
\usage{
blue_cliff_record
}
\description{
A koan collection written in 1125 and compiled by by Zen Master Yuanwu Keqin
that includes the 100 Verses on Old Cases, a compilation of 100 koans
collected by Xuedou Chongxian (Setcho), whose commentary appear ocassionally
throughout.
}
\keyword{datasets}
