% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koan.R
\name{collect_koans}
\alias{collect_koans}
\title{Bind koan collections together}
\usage{
collect_koans()
}
\value{
a \code{tbl.} containing all 4 koan collections
}
\description{
This is a convenience function that binds the koan data together into a
single \code{tbl.}
}
\examples{
library(dplyr)

collect_koans() \%>\% filter(type == "main_case")
}
