% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gateless_gate}
\alias{gateless_gate}
\title{Record of the Transmission of the Light}
\format{A data frame with four variables: \code{collection}, \code{case},
  \code{type}, and \code{text}. \code{collection} is the title of the text,
  "Record of the Transmission of the Light". \code{case} indicates the koan
  case number. \code{type} indicates the text type, one of: "title", the
  title of the case, "main_case", the main case of the koan, "commentary",
  commentary by Keizan Zenji, or "capping_verse", a short verse by Keizan
  Zenji summarizing the essence of the koan.}
\source{
https://shastaabbey.org/publications/
}
\usage{
gateless_gate
}
\description{
A koan collection written in 1300 by Zen Master Keizan Jokin that describes
the enlightenment experience of of the Buddha and the 53 ancestors that led
to Keizan Zenki's school of Zen.
}
\keyword{datasets}
