% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koan.R
\name{Sample a koan}
\alias{Sample a koan}
\alias{koan}
\alias{title}
\alias{main_case}
\alias{commentary}
\alias{capping_verse}
\title{Sample a koan}
\usage{
koan(collection = NULL, case = NULL, verbose = TRUE)

title(collection = NULL, case = NULL, verbose = TRUE)

main_case(collection = NULL, case = NULL, verbose = TRUE)

commentary(collection = NULL, case = NULL, verbose = TRUE)

capping_verse(collection = NULL, case = NULL, verbose = TRUE)
}
\arguments{
\item{collection}{a \code{data.frame}, the koan collection to sample from.
Default is \code{NULL}, which randomly selects a collection to pick from.}

\item{case}{the case number. Default is \code{NULL}, which randomly selects a
case.}

\item{verbose}{logical. Should koan be printed to console?}
}
\value{
verboses a koan (or part of it) and invisibly returns either a \code{tbl}
  containing the text (\code{koan()}) or a character vector.
}
\description{
Sample a koan
}
\examples{
koan(gateless_gate, 1)

}
