% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{book_of_equanimity}
\alias{book_of_equanimity}
\title{Book of Equanimity}
\format{A \code{tbl} with four variables: \code{collection}, \code{case},
  \code{type}, and \code{text}. \code{collection} is the title of the text,
  "Book of Equanimity". \code{case} indicates the koan case number.
  \code{type} indicates the text type, one of: "title", the title of the
  case, or "main_case", the main case of the koan.}
\source{
http://www.thezensite.com/ZenTeachings/KoanStudies/Shoyoroku.pdf
}
\usage{
book_of_equanimity
}
\description{
A koan collection written in 1224 by Zen Master Wansong Xingxiu that is
comprised of 100 koans written by 11th century Zen Master Hongzhi Zhengjue
}
\keyword{datasets}
