% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_koan.R
\name{Fake Koans}
\alias{Fake Koans}
\alias{fake_koan}
\alias{fake_title}
\alias{fake_main_case}
\alias{fake_commentary}
\alias{fake_capping_verse}
\title{Generate fake koans using a Markov model}
\usage{
fake_koan(verbose = TRUE)

fake_title(verbose = TRUE)

fake_main_case(n_lines = sample(1:20, 1), verbose = TRUE)

fake_commentary(n_lines = sample(1:20, 1), verbose = TRUE)

fake_capping_verse(n_lines = sample(3:6, 1), verbose = TRUE)
}
\arguments{
\item{verbose}{logical. Should koan be printed to console?}

\item{n_lines}{The amount of lines to create. Default randomly samples a
number.}
}
\value{
verboses a koan (or part of it) and invisibly returns a \code{tbl}
  containing the text.
}
\description{
Generate fake koans using a Markov model
}
\examples{
fake_koan()
}
