% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gateless_gate}
\alias{gateless_gate}
\title{The Gateless Gate}
\format{A \code{tbl} with four variables: \code{collection}, \code{case},
  \code{type}, and \code{text}. \code{collection} is the title of the text,
  "The Gateless Gate". \code{case} indicates the koan case number.
  \code{type} indicates the text type, one of: "title", the title of the
  case, "main_case", the main case of the koan, "commentary", commentary by
  Wumen (Mumon), or "capping_verse", a short verse by Wumen (Mumon)
  summarizing the essence of the koan.}
\source{
http://info.stiltij.nl/publiek/meditatie/koan/mumonkan-shimomisse.pdf
}
\usage{
gateless_gate
}
\description{
A koan collection written in 1228 by Zen Master Wumen Huikai (Japanese: Mumon
Ekai) that includes a commentary and capping verse by Wumen for each koan.
}
\keyword{datasets}
